iOS: finish x86 asm, per-file VPC conditions, port SDL manager, misc fixes
- arm64: ThreadPause uses yield, vertex copies use memcpy, cpuid is
  emulated to report the MMX/SSE/SSE2/SSE3 that sse2neon provides, fast
  clock reads cntvct_el0, VectorNormalizeFast has a C path.
- VPC parser: a $File statement listing several files keeps each file's
  own [condition] (Windows-only audio files were being built).
- appframework: use the source-engine port's SDL launcher manager (iOS
  window, EGL/ANGLE context, Metal layer) adapted to CS:GO's ILauncherMgr;
  SetGammaRamp and window ref counting are part of the interface on iOS.
- togl types: 32-bit ULONG/LONG/HRESULT as CS:GO's togl had.
- .nut scripts are embedded as <name>_nut.h at build time (Valve's
  texttoarray.pl); stdshader_dbg finds its generated shader headers.
- Restore VERSION_SAFE_STEAM_API_INTERFACES and USE_BREAKPAD_HANDLER: the
  engine uses the declarations they enable.
- Drop CS:GO's Linux-configured FreeType header copy; use the real one.
- iOS takes the OpenAL voice recorder, Darwin statfs, SDL message boxes,
  and skips macOS system proxies, Carbon WindowRef/cursor calls, the
  32-bit-inode check, FSEvents and system().
- Small fixes: pointer-vs-zero comparisons, C99 round() clash, Valve-tree
  include paths, FCVAR_INTERNAL_USE, dbg.h SDL assert parent, vphysics
  save/restore macro and collision set include.

+++ public/togl/dxabstract_types.h
@@ -90,14 +90,15 @@ typedef void* VD3DHANDLE;
 #if !defined(_WINNT_)
 
 	typedef int INT;
-	typedef unsigned long ULONG;
-	typedef long LONG;
+	// 32-bit as in CS:GO's togl: engine and shaderapi code assume it on 64-bit
+	typedef unsigned int ULONG;
+	typedef int LONG;
 	typedef float FLOAT;
 	typedef unsigned int DWORD;
 	typedef unsigned short WORD;
 	typedef long long LONGLONG;
 	typedef unsigned int UINT;
-	typedef long HRESULT;
+	typedef int HRESULT;
 	typedef unsigned char BYTE;
 	#define CONST const
